cnn_interface_sj/ApplicationFormClassification/requirements.txt
absl-py==0.8.0
astor==0.8.0
backports.functools-lru-cache==1.5
certifi==2019.11.28
chardet==3.0.4
cheroot==6.5.7
cycler==0.10.0
Cython==0.28.2
gast==0.2.2
grpcio==1.23.0
h5py==2.9.0
idna==2.8
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
matplotlib==2.2.2
more-itertools==7.2.0
numpy==1.18.0
opencv-python==3.4.0.14
Pillow==6.2.1
protobuf==3.9.1
pyasn1==0.4.7
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
scipy==1.3.1
six==1.12.0
tensorboard==1.9.0
torch==1.1.0
torchsummary==1.5.1
torchvision==0.2.0
urllib3==1.25.7
web.py==0.40.dev1
Werkzeug==0.15.5
# python3.6.2
